refactor: small fixes to the current implementation
