+++ src/components/Navbar.jsx
@@ -1,4 +1,3 @@
-"use client"
 import classNames from 'classnames';
 import { useLocation } from "react-router-dom";
 import { Link } from "react-router-dom";
