+++ src/pages/DestinationPage/AddWishlistItem.jsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { useState } from 'react';
 import styles from './DestinationPage.module.css';
