+++ src/pages/NasaCollaborationPage/NasaCollaborationPage.jsx
@@ -1,5 +1,3 @@
-"use client";
-
 import React, { useState, useEffect } from 'react';
 import styles from './NasaCollaborationPage.module.css';
